Revert "Implelment safety and regard metrics in pure unitxt (#983)"
This reverts commit 0a9448807844240407c50640809bcda4efdc1fb0.
Reason for the revert: This commmit added the class Safety. Same class is already implemented in fm-eval. Since this class is registered, we can havetwo classes with the same name. We should move the implementation for fm-eval to unitxt.
