+++ src/unitxt/catalog/metrics/safety.json
@@ -1,5 +1,9 @@
 {
-    "__type__": "safety",
+    "__type__": "huggingface_metric",
+    "hf_metric_name": "src/metrics/safety",
+    "main_score": "safety",
+    "hf_main_score": "score",
+    "scale": 1.0,
     "n_resamples": null,
     "prediction_type": "Any"
 }
